In summary this builds a Windows .EXE on Linux and what it builds is an ultra-low-latency X11 server. I.e. Lightning fast.
1. New build system for x86_64-linux-gnu-gcc in Makefile.mingw64.
2. Fix many build warnings.
3. Create new pthread library pthreadlight.c which has only what X11 needs.
4. Strip openssl library which was only being used for SHA1.
5. MsgWaitForMultipleObjects implementation to replace select.
6. WSARecv to handle iovec.
7. Socket read-ahead to reduce kernel ops.
8. Various fixes for latency improvements.
9. New install/ming64x/fonts/ directory with all fonts pre-built.
10. Many WIN32 compatibility bug fixes.

+++ xorg-server/hw/xwin/winengine.c
@@ -173,7 +173,7 @@ winSetEngine(ScreenPtr pScreen)
             winSetEngineFunctionsShadowDDNL(pScreen);
             break;
         default:
-            FatalError ("winSetEngine - Invalid engine type %d\n",pScreenInfo->dwEngine);
+            FatalError ("winSetEngine - Invalid engine type %d\n",(int)pScreenInfo->dwEngine);
         }
         return TRUE;
     }
