+++ xorg-server/xkb/xkb.c
@@ -2459,7 +2459,7 @@ _XkbSetMapCheckLength(xkbSetMapReq *req)
     if (len == req_len)
         return Success;
 bad:
-    ErrorF("[xkb] BOGUS LENGTH in SetMap: expected %zd got %zd\n", len, req_len);
+    ErrorF("[xkb] BOGUS LENGTH in SetMap: expected %ld got %ld\n", (long)len, (long)req_len);
     return BadLength;
 }
